pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anti-aging-scraper"
version = "1.0.0"
description = "Anti-aging research article scraper"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.25.1",
    "feedparser>=6.0.0",
    "pandas>=1.3.0",
    "beautifulsoup4>=4.9.3",
    "newspaper3k>=0.2.8",
    "lxml[html_clean]>=4.6.0",
    "schedule>=1.1.0",
    "openai>=0.27.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.8",
] 